requirements.txt
# SlopeGuard - Requirements
# Python 3.9+

# Core
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Geospatial
rasterio>=1.2.0
geopandas>=0.10.0
shapely>=1.8.0
pyproj>=3.2.0
fiona>=1.8.0
h5py>=3.6.0

# SAR/InSAR Processing
asf_search>=6.0.0
sentinelsat>=1.1.0
mintpy>=1.5.0

# Machine Learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
folium>=0.12.0
plotly>=5.5.0
branca>=0.6.0

# API
fastapi>=0.70.0
uvicorn>=0.15.0
pydantic>=1.9.0

# Weather/HTTP
requests>=2.26.0
httpx>=0.23.0

# Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
tqdm>=4.62.0
click>=8.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# AI Agent
google-genai>=1.0.0

# Testing
pytest>=7.0.0
